läuft mit 1GB stick (FAT16) und real PC
git-svn-id: svn://svn.code.sf.net/p/prettyos/code/trunk/Source@420 31957799-4632-4615-9793-72db6d48bcac

+++ kernel/include/usb2.h
@@ -119,7 +119,7 @@ uint8_t usbTransferGetConfiguration(uint32_t device);
 void usbTransferBulkOnlyMassStorageReset(uint32_t device, uint8_t numInterface);
 uint8_t usbTransferBulkOnlyGetMaxLUN(uint32_t device, uint8_t numInterface);
 
-void usbSendSCSIcmd(uint32_t device, uint32_t endpointOut, uint32_t endpointIn, uint8_t SCSIcommand, uint32_t LBA, uint32_t TransferLength, bool MSDStatus);
+void usbSendSCSIcmd(uint32_t device, uint32_t endpointOut, uint32_t endpointIn, uint8_t SCSIcommand, uint32_t LBA, uint16_t TransferLength, bool MSDStatus);
 
 void addDevice(struct usb2_deviceDescriptor* d, usb2_Device_t* usbDev);
 void showDevice(usb2_Device_t* usbDev);
